Price modification and bugfixes
Added price modifier and fixed major bugs with UPC barcode length
checking. Also nulls price if the price is not available but the item is
found. Changed the close icon to a material icons and added the price
modifier icon as a material icon, too. Future-proofed my alignment code.
Italicized close icon and other icons. Removed padding on icons because
it caused a problem with multiple items.